requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.5
google-auth==1.31.0
requests==2.26.0
pyrebase
PyJWT